stream-input: introduce INPUT-STREAM-KERNEL class
For easier review it has been introduced as a separate commit. Classes
STANDARD-INPUT-STREAM and STANDARD-EXTENDED-INPUT-STREAM will be build on top
of the INPUT-STREAM-KERNEL. Parts of other stream protocols are implemented
for this base class.

Extended Input Stream Protocol:

- stream-input-wait
- stream-listen
- stream-process-gesture
- stream-read-gesture
- stream-unread-gesture
- stream-input-buffer, (setf stream-input-buffer)

Basic Input Stream Protocol:

- stream-clear-input

Input Editing Stream Protocol:

- stream-process-gesture

Additionally it introduces its own protocol functions:

- stream-append-gesture
- stream-gesture-available-p

INPUT-STREAM-KERNEL class is meant to implement interactions between the event
queue (which is the sheet input protocol domain) and the input buffer (which
is the input stream domain). This class is not directly mentioned in the spec
but is hinted in both STANDARD-INPUT-STREAM and STANDARD-EXTENDED-INPUT-STREAM
class descriptions.

The gist of its inner workings is as follows:

- STREAM-INPUT-WAIT operation calls event-listen-or-wait and handles events
  available in the queue until INPUT-BUFFER has available input. STREAM-LISTEN
  is just a trampoline to STREAM-INPUT-WAIT.

- all other methods which need access to stream gestures operate directly on
  the INPUT-BUFFER. STREAM-READ-GESTURE invokes STREAM-INPUT-WAIT.